gpu: ion: Limit scope of local functions
Several functions that should not be used outside of
the ion.c file does not have static keyword. In
addition two of the functions should not be exported.

Add static keyword to functions that are local to the
translation unit and remove EXPORT_SYMBOL declarations
that are not needed.

Signed-off-by: Olav Haugan <ohaugan@codeaurora.org>
(cherry picked from commit 0fa9b60cc8a91b20f4c335f7bac844867506885b)

Change-Id: I1b3344cbc191dae9bc3126fbc1bcd381a3dc1ad1
Signed-off-by: Yen-Pin Hsiao <yphsiao@codeaurora.org>